Adjust config options for use with both stock LCD and touchscreen
Moved LED configuration to a separate folder in options/led.

Moved LCD led control into options/led/lcd.cfg

Moved the macro for mainled into this folder so the correct macro is defined based on whether it is defined as a pin or led.

Removed the hdmi folder for now as it is empty. Can add it back if more configuration for the hdmi option is added.